Ulf Andrae & Trygve Aspelien: LSELECT functionality
LSELECT functionality has changed in SURFEX 8.1. Now a CSELECT array is passed as an argument to all the write_surf calls and the global variables LSELECT/CSELECT are gone after the restructuring. However, in write_surf_atmn some records (like version) must always be written. Since the test is now always done on CSELECT this commit modifies the update after reading the namelist to take this into account and add the the extra fields if needed.