Trygve Aspelien & Mariken Homleid: snow assimilation
- Adapt snow assimilation to be able to be used for ISBA-ES
Rewriting the snow update and adapt some other routines for snow initialization and sanity checks to be able to handle multiple layers
- Reformulate the weighting of the observed snow.
The old approach had large weaknesses if the difference in SWE was too large between the patches, but did conserve the grid value.
The new approach assumes that the observed value is representable for patch 1 (open land), and the difference between the SWE values
at the different patches is used as a weighting function with some constraints keeping snow on the patches if patch 1 has snow and
make sure the patches do not get too much snow if patch 1 has a lower SWE value. The new approach will not conserve the grid value
compared to the observed value.
- Bring in the CARRA features of user initializing (update) of snow and keep snow where SURFEX has a large enough fraction of permanent snow

LOGICAL                               :: LSWEPSINI            ! Logical switch to set initial values of SWE on
                                                              ! vegetation type PERMANENT SNOW in the snow analysis
REAL                                  :: XSWEPSINI            ! Initial value of SWE on PERMANENT SNOW

LOGICAL                               :: LSWEPSMIN            ! Logical switch to set a lower limit for SWE on
                                                              ! vegetation type PERMANENT SNOW in the snow analysis
REAL                                  :: XSWEPSMIN            ! Lower limit of SWE on PERMANENT SNOW

Mariken Homleid: Re-introduce old weighting, but with limitations, when using first guess from grid average snow (AKA CANARI)