Yurii Batrak: SICE
Fix incorrect MPI communication in ol_propagate_ice

Provide parallelised implementation of SICE propagation routines

Originally when Canari (or Soda) is run in a parallel environment all
the manipulations with the SICE model fields were performed on a single
processor with the results sent to all the processors. This approach
helps to simplify the propagation routines which require access to
the full SICE fields at some stages. Although, for bigger domains
with considerable sea ice cover it causes a noticeable slowdown
due to the inefficient load balance. The updated procedures do not
rely on a single processor to do all the work but redistribute the load
for all the processors.

Ulf Andrae: bugfix: Fix misplaced deallocation in SICE