Yurii Batrak: Provide transfer coefficients for the turbulent flux of water over sea ice
Over sea ice-covered grid cells transfer coefficients for the turbulent
flux of water are not calculated by the ICE_SEA_FLUX routine, but the
calling routine COUPLING_SEAFLUX_n expects them to be available. During
the initialization stage the corresponding variable is filled with
XUNDEF but it is never replaced by the actual values. As a result when
calculating a grid cell average value of this coefficient XUNDEF is used
for the ice-covered part of a grid cell resulting in unphysical values.
To alleviate this issue transfer coefficients for the turbulent flux of
water are set to be equal to these for heat.