Ekaterina Kurzeneva: Added functionality of limiting the land/water(sea+lake) fractions.
1. Added functionality of limiting the land/water(sea+lake) fractions
for some (small) percentage value. This functionality is needed for
the correct snow DA. Limiting is applied on the level of both covers
and "sea/water" fractions within covers.

2. Added functionality to define the limit for the land/water(sea+lake)
fractions in the namelist: XRM_WM in NAM_COVER namelist of SURFEX.

3. Reading of the NAM_COVER from ini_data_cover, because XRM_WM is
needed already there. For this, moving of formal parameters from
read_nam_pgd_cover to global variables in modd_data_cover_par,
as well as changing of some calls.

Item 1 is essential, while technical details from items 1 and 2 may be
changed in future.