Fix LMEB_LITTER problem (Hirlam/Harmonie#1160)
Allow LMEB_LITTER to be activated in general but to be finally decided per vegtype/patch by the PGNDLITTER litter thickness given in ini_data_param.F90. So, if PGNDLITTER is set to 0 the litter parameterisation will not be used for this vegtype.
Also, allow MEB to be activated for any patch. Until now only forest patches were allowed.